new debounce tester "cpp_test" (faster+html output)
`cd tools/debounce_test/cpp_test; make -j6 test; firefox ...`

Compared to the original tester, it's faster (2 seconds instead of 40),
run multiple tests for each input data file (samples nearest(s) and
average(s)), outputs as dynamic html/js table.

But debugging output is less advanced/readable. Outputs debug on stderr
with -d, example:

`./debounce-integrator -d ../testcases/chatterboard/key-b--10-presses-fast.data 2> debug.txt`

+++ tools/debounce_test/debounce_test.c
@@ -45,7 +45,7 @@ void get_input () {
             comments[scan_counter+1][79]= '\0';
             continue;
         }
-        for(uint16_t i = 0; (i< sizeof(buffer) && buffer[i+1] != 0)  ; i++) {
+        for(int i = 0; (i < (int)sizeof(buffer) && buffer[i+1] != 0)  ; i++) {
             if(buffer[i] == '#') {
                 break;
             } else if(buffer[i] == '0') {
